@@ -22,14 +22,15 @@ export class SelectDropdownComponent implements OnInit {
 
   constructor() { }
   @Input('teamMembers') teamMembers: string[];
-  @Output() setteamMember = new EventEmitter<{ teamMember: string }>();
+  @Output() setTeamMember = new EventEmitter<{ teamMember: string }>();
   selectedOption: string = "Team";
 
   ngOnInit(): void {
+    console.log(this.teamMembers);
   }
 
   setSelectedOption() {
-    this.setteamMember.emit({ teamMember: this.selectedOption });
+    this.setTeamMember.emit({ teamMember: this.selectedOption });
   }
 
 }
