@@ -10,6 +10,7 @@ import { CdkDragDrop, moveItemInArray, transferArrayItem } from '@angular/cdk/dr
 import { ToolsService } from 'src/app/services/tool/tools.service';
 import { AuthService } from 'src/app/services/auth.service';
 import { Sprint } from 'src/app/Interface/TeamInterface';
+import { CookieService } from 'ngx-cookie-service';
 
 @Component({
   selector: 'app-kanban-board',
@@ -33,8 +34,11 @@ export class KanbanBoardComponent implements OnInit {
   currentSprintName: string;
   currentSprintNumber: number;
   filterSprintNumber: number;
+  currentStatusLabels: string[];
+  currentSelectedStatusLabels: string;
+  allTasks : Tasks[];
 
-  constructor(public navbarHandlerService: NavbarHandlerService, public startService: StartServiceService, public applicationSettingsService: ApplicationSettingsService, public backendService: BackendService, private functions: AngularFireFunctions, public errorHandlerService: ErrorHandlerService, public toolsService: ToolsService, public authService: AuthService) { }
+  constructor(public navbarHandlerService: NavbarHandlerService, public startService: StartServiceService, public applicationSettingsService: ApplicationSettingsService, public backendService: BackendService, private functions: AngularFireFunctions, public errorHandlerService: ErrorHandlerService, public toolsService: ToolsService, public authService: AuthService, public cookieService: CookieService) { }
 
   ngOnInit(): void {
     this.navbarHandlerService.resetNavbar();
@@ -63,9 +67,9 @@ export class KanbanBoardComponent implements OnInit {
   }
 
   readData() {
-    this.selectedStatusLabels = ['Ice Box', 'Ready to start', 'Under Progress', 'Blocked'];
     this.selectedTeamId = this.startService.selectedTeamId;
     this.statusLabels = this.applicationSettingsService.status;
+    this.selectedStatusLabels = this.statusLabels.slice(0,4);
     this.currentSprintNumber = this.startService.currentSprintNumber;
     if(this.currentSprintNumber==-1){
       this.currentSprintName= "Backlog";
@@ -101,8 +105,17 @@ export class KanbanBoardComponent implements OnInit {
           this.errorHandlerService.getErrorCode(this.componentName, "InternalError","Api");
           console.error(error);
         },
-        complete: () => console.info('Successful updated Selected Team in db')
+        complete: (() => {
+          this.cookieService.set("userSelectedTeamId", teamId);
+          console.info('Successful updated Selected Team in db');
+        })
     });
+  }
+
+  changeStatusLabels(newStatus, existingStatus){
+    let x = this.selectedStatusLabels.indexOf(existingStatus)
+    this.selectedStatusLabels[x] = newStatus;
+    this.putDataInTasksArray(this.allTasks);
   }
 
   changeSprintNumber() {
@@ -134,6 +147,7 @@ export class KanbanBoardComponent implements OnInit {
     callable({OrgDomain: orgDomain, TeamId: this.selectedTeamId, SprintNumber: this.currentSprintNumber }).subscribe ({
       next: (data) => {
         this.putDataInTasksArray(data.data);
+        this.allTasks = data.data;
         this.showLoader = false;
         console.log("read tasks successfully!")
       },
@@ -188,18 +202,5 @@ export class KanbanBoardComponent implements OnInit {
     }
   }
 
-  getColor(status: string) {
-    if (status == 'Ready to start') {
-      return '#673AB7';
-    } else if (status == "Ice Box") {
-      return '#42A5F5';
-    } else if (status == "Under Progress") {
-      return '#FC6A03';
-    } else if (status == "Blocked") {
-       return '#E64336';
-    } else {
-      return '#00F11E';
-    }
-  }
 
 }
