@@ -13,10 +13,8 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  * See the MIT License for more details.
  ***********************************************************/
-import { newArray } from '@angular/compiler/src/util';
 import { Component, OnInit } from '@angular/core';
 import { AngularFireFunctions } from '@angular/fire/compat/functions';
-import { Router } from '@angular/router';
 import { ErrorHandlerService } from 'src/app/services/error-handler/error-handler.service';
 import { map } from 'rxjs';
 import { Post } from 'src/app/Interface/SocialInterface';
@@ -31,47 +29,57 @@ import { UserServiceService } from 'src/app/services/user-service/user-service.s
   styleUrls: ['./social-page.component.css']
 })
 export class SocialPageComponent implements OnInit {
-
+  
   componentName: string = "SOCIAL-PAGE"
   showloader: boolean = false
   dataReady: boolean = false
   createPostEnabled: boolean = false
   public posts: Post[]
-  public recentPosts: Post[]
+  public recentPosts: Post[] = []
+
+  pageReady:boolean = false;
 
   PostId: string
 
-  constructor(private router: Router, private navbarHandler: NavbarHandlerService, private functions: AngularFireFunctions, public errorHandlerService: ErrorHandlerService, public popupHandlerService: PopupHandlerService, public userService:UserServiceService, public authService: AuthService) { }
+  constructor(private navbarHandler: NavbarHandlerService, private functions: AngularFireFunctions, public errorHandlerService: ErrorHandlerService, public popupHandlerService: PopupHandlerService, public userService:UserServiceService, public authService: AuthService) { }
 
   ngOnInit(): void {
     this.navbarHandler.resetNavbar();
-    this.loadSocialPageData();
+    this.authService.getUserSettings();
+    this.authService.userAppSettingObservable.subscribe((data)=>{
+      this.pageReady = true;
+      this.loadSocialPageData();
+    });
   }
 
   loadSocialPageData() {
+    this.showloader = true;
     const callable = this.functions.httpsCallable("socialPage/getAllPosts");
     callable({}).pipe(map(res=>{
       const data = res.data as Post[];
       return data
     })).subscribe({
       next:(data)=>{
-      if(data) {
-        this.posts = data;
-        this.posts.forEach(element => {
-          this.userService.checkAndAddToUsersUsingUid(element.Uid);
-        });
-        this.userService.fetchUserDataUsingUID().subscribe(()=>{
-          this.dataReady = true;
-        });
+        if(data) {
+          this.posts = data;
+          this.posts.forEach(element => {
+            this.userService.checkAndAddToUsersUsingUid(element.Uid);
+          });
+          this.userService.fetchUserDataUsingUID().subscribe(()=>{
+            this.dataReady = true;
+          });
+          this.loadRecentActivity();
+        }
+      },
+      error:(error)=>{
+        this.errorHandlerService.showError = true;
+        this.errorHandlerService.getErrorCode(this.componentName, "InternalError","Api");
+        console.error(error);
+      },
+      complete:()=>{
+        this.showloader = false;
       }
-      this.showloader = false
-      this.loadRecentActivity();
-    },
-    error:(error)=>{
-      this.errorHandlerService.showError = true;
-      this.errorHandlerService.getErrorCode(this.componentName, "InternalError","Api");
-      console.error(error);
-    } });
+    });
   }
 
   createPost() {
@@ -86,7 +94,7 @@ export class SocialPageComponent implements OnInit {
 
   loadRecentActivity(){
     const newarray = this.posts.filter((data)=>{
-      if(data.Uid == this.authService.userAppSetting.uid) {
+      if(this.authService.userAppSetting != undefined && data.Uid == this.authService.userAppSetting.uid) {
         return data;
       }
     });
