@@ -20,6 +20,7 @@ import { Observable } from 'rxjs';
 import { ThemeService } from './theme/theme.service';
 import { map } from 'rxjs/operators';
 import { CookieService } from 'ngx-cookie-service';
+import { FileData } from '../Interface/FileInterface';
 
 @Injectable({
   providedIn: 'root',
@@ -41,14 +42,17 @@ export class AuthService {
   public userName:string = "";
   public landingToSocial: boolean = false;
 
-  educations: MyEducationData;
-  public educationCollectionData: Observable<MyEducationData>
+  educations: MyEducationData[];
+  public educationCollectionData: Observable<MyEducationData[]>
 
-  experiences: MyExperienceData;
-  public experienceCollectionData: Observable<MyExperienceData>
+  experiences: MyExperienceData[];
+  public experienceCollectionData: Observable<MyExperienceData[]>
 
-  projects: MyProjectData;
-  public projectCollectionData: Observable<MyProjectData>
+  projects: MyProjectData[];
+  public projectCollectionData: Observable<MyProjectData[]>
+
+  filesData: FileData[];
+  public filesCollectionData: Observable<FileData[]>
 
   constructor(private cookieService: CookieService, public afauth: AngularFireAuth, private functions: AngularFireFunctions, public themeService: ThemeService) { }
 
@@ -150,7 +154,7 @@ export class AuthService {
     const callable = this.functions.httpsCallable("users/getAllEducation");
     this.educationCollectionData = callable({Uid: uid }).pipe(
       map(actions => {
-        this.educations = actions.data as MyEducationData;
+        this.educations = actions.data as MyEducationData[];
         return this.educations;
       }));
       return this.educationCollectionData;
@@ -160,7 +164,7 @@ export class AuthService {
     const callable = this.functions.httpsCallable("users/getAllExperience");
     this.experienceCollectionData = callable({Uid: uid }).pipe(
       map(actions => {
-        this.experiences = actions.data as MyExperienceData;
+        this.experiences = actions.data as MyExperienceData[];
         return this.experiences;
       }));
       return this.experienceCollectionData;
@@ -170,10 +174,20 @@ export class AuthService {
     const callable = this.functions.httpsCallable("users/getAllProject");
     this.projectCollectionData = callable({Uid: uid }).pipe(
       map(actions => {
-        this.projects = actions.data as MyProjectData;
+        this.projects = actions.data as MyProjectData[];
         return this.projects;
       }));
       return this.projectCollectionData;
+  }
+
+  getUserProfilePic(uid: string) {
+    const callable = this.functions.httpsCallable("librarian/getFilesInUser");
+    this.filesCollectionData = callable({UID: uid }).pipe(
+      map(actions => {
+        this.filesData = actions.data as FileData[];
+        return this.filesData;
+      }));
+      return this.filesCollectionData;
   }
 
   getAppKey() {
