@@ -29,7 +29,9 @@ export class PopupHandlerService {
   quickNotesDescription: string = ""
   addLabelEnabled: boolean = false
   milestoneId:string = ""
-  
+  addMilestoneActive:boolean = true;
+  addTaskActive:boolean = true;
+  requestDemoEnabled:boolean = false;
   constructor() { }
 
   resetPopUps() {
